requirements.txt
youtube_dl==2021.12.17
google-api-python-client==2.35.0
python-dotenv==0.14.0
google-auth == 1.32.1